+++ src/ShowKeys/ErrorEntry.cs
@@ -1,7 +1,0 @@
-﻿namespace ShowKeys
-{
-    // Represents an error when key combo couldn't be determined
-    public class ErrorEntry : KeyEntryItem
-    {
-    }
-}