+++ src/CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Initial release of ShowKeys application
-- Transparent overlay displaying the last 15 keyboard combinations
+- Transparent overlay displaying the last N keyboard combinations
 - Support for modifier keys (Ctrl, Alt, Shift, Win) combined with other keys
 - Display of special keys (Enter, Tab, Backspace, etc.)
 - Positioning above the taskbar
 - Some padding is added after 1.5 seconds to separate fast key sequences
+- System tray icon menu item: "Quit" exits the applicatio
+- System tray icon menu item: "Settings" opens application settings
+- Settings: # of kb combinations is now configurable
 
 ### Changed